+++ lib/brando/images/image.ex
@@ -14,18 +14,6 @@ defmodule Brando.Image do
   @required_fields ~w(image image_series_id)a
   @optional_fields ~w(sequence creator_id)a
 
-  @derive {Poison.Encoder,
-           only: [
-             :id,
-             :image,
-             :creator,
-             :creator_id,
-             :image_series_id,
-             :image_series,
-             :sequence,
-             :inserted_at,
-             :updated_at
-           ]}
   @derive {Jason.Encoder,
            only: [
              :id,