+++ lib/brando/pages/page_fragment.ex
@@ -7,7 +7,6 @@ defmodule Brando.Pages.PageFragment do
 
   use Brando.Web, :schema
   use Brando.Villain, :schema
-  alias Brando.Type.Json
 
   @required_fields ~w(parent_key key language data creator_id)a
   @optional_fields ~w(html)a
@@ -21,14 +20,13 @@ defmodule Brando.Pages.PageFragment do
     creator_id
   )a
 
-  @derive {Poison.Encoder, only: @derived_fields}
   @derive {Jason.Encoder, only: @derived_fields}
 
   schema "pagefragments" do
     field :parent_key, :string
     field :key, :string
     field :language, :string
-    field :data, Json
+    field :data, :map
     field :html, :string
     belongs_to :creator, Brando.User
     timestamps()
@@ -61,7 +59,7 @@ defmodule Brando.Pages.PageFragment do
 
   def encode_data(params) do
     if is_list(params.data) do
-      Map.put(params, :data, Poison.encode!(params.data))
+      Map.put(params, :data, Jason.encode!(params.data))
     else
       params
     end